+++ src/input/zapper.c
@@ -28,7 +28,7 @@
 #define MAX_TOLERANCE 20
 static uint32 targetExpansion[MAX_TOLERANCE+1];
 #endif
-static int tolerance;
+static uint32 tolerance;
 
 typedef struct {
 	uint32 mzx, mzy, mzb;
@@ -82,7 +82,7 @@ static void FP_FASTAPASS(3) ZapperFrapper(int w, uint8 * bg, uint8 * spr, uint32
 
 				sum = palo[a1].r + palo[a1].g + palo[a1].b;
 				if (sum >= 100 * 3) {
-					ZD[w].zaphit = ((uint64)linets + (xs + 16) * (PAL ? 15 : 16)) / 48 + timestampbase;
+					ZD[w].zaphit = ((uint64)linets + (uint64)(xs + 16) * (PAL ? 15 : 16)) / 48 + timestampbase;
 					goto endo;
 				}
 			}
@@ -167,7 +167,7 @@ static uint32 InefficientSqrt(uint32 z) {
 void FCEU_ZapperSetTolerance(int t)
 {
 #ifdef ROUNDED_TARGET
-	uint32_t y;
+	uint32 y;
 	tolerance = t <= MAX_TOLERANCE ? t : MAX_TOLERANCE;
 	for (y = 0; y <= tolerance; y++)
 		targetExpansion[y] = InefficientSqrt(tolerance*tolerance-y*y);
